requirements.txt
streamlit
pydantic
langchain
anthropic
python-dotenv
instructor
marvin
uuid
fix-busted-json